Creates new pending broker controller
