+++ app/helpers/pending_brokers_helper.rb
@@ -1,0 +1,2 @@
+module PendingBrokersHelper
+end
