+++ app/views/layouts/_header.html.erb
@@ -11,17 +11,17 @@
             <% if current_page?(users_admin_index_path) %>
               <li><%= link_to "All Users", users_admin_index_path, class: 'btn btn-outline-secondary mx-2 text-decoration-none active' %></li>
               <li class='mx-2 position-relative'>
-                <%= link_to "Pending Brokers", brokers_path, class: 'btn btn-outline-secondary text-decoration-none'%>
+                <%= link_to "Pending Brokers", pending_brokers_path, class: 'btn btn-outline-secondary text-decoration-none'%>
                 <% if !@unapproved_users.empty? %>
                   <span class='position-absolute top-0 start-100 translate-middle badge rounded-pill bg-danger'>
                     <%= @unapproved_users.count %>
                   </span>
                 <% end %>
               </li>
-            <% elsif current_page?(brokers_path) %>
+            <% elsif current_page?(pending_brokers_path) %>
               <li><%= link_to "All Users", users_admin_index_path, class: 'btn btn-outline-secondary mx-2 text-decoration-none' %></li>
               <li class='mx-2 position-relative'>
-                <%= link_to "Pending Brokers", brokers_path, class: 'btn btn-outline-secondary text-decoration-none active' %>
+                <%= link_to "Pending Brokers", pending_brokers_path, class: 'btn btn-outline-secondary text-decoration-none active' %>
                 <% if !@unapproved_users.empty? %>
                   <span class='position-absolute top-0 start-100 translate-middle badge rounded-pill bg-danger'>
                     <%= @unapproved_users.count %>
@@ -31,7 +31,7 @@
             <% else %>
               <li><%= link_to "All Users", users_admin_index_path, class: 'btn btn-outline-secondary mx-2 text-decoration-none' %></li>
               <li class='mx-2 position-relative'>
-                <%= link_to "Pending Brokers", brokers_path, class: 'btn btn-outline-secondary text-decoration-none'%>
+                <%= link_to "Pending Brokers", pending_brokers_path, class: 'btn btn-outline-secondary text-decoration-none'%>
                 <% if !@unapproved_users.empty? %>
                   <span class='position-absolute top-0 start-100 translate-middle badge rounded-pill bg-danger'>
                     <%= @unapproved_users.count %>
